railway-deploy/florida-data-ingestion/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.24.1
supabase==2.0.2
beautifulsoup4==4.12.2
lxml==4.9.3
sentry-sdk==1.38.0
openai==1.3.7
transformers==4.35.2
torch==2.1.1
pandas==2.1.3
geopandas==0.14.1
playwright==1.40.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
